@@ -18,24 +18,28 @@
     }
   ],
   "require": {
-    "php": "~7.4.0 || ~8.0.0 || ~8.1.0",
+    "php": "~8.1.0 || ~8.2.0",
     "ext-json": "*",
     "ext-mbstring": "*",
     "phpoffice/phpspreadsheet": "^1.23",
-    "typo3/cms-backend": "^10.4 || ^11.5",
-    "typo3/cms-core": "^10.4 || ^11.5",
-    "typo3/cms-extbase": "^10.4 || ^11.5",
-    "typo3/cms-fluid": "^10.4 || ^11.5",
-    "typo3/cms-fluid-styled-content": "^10.4 || ^11.5",
-    "typo3/cms-frontend": "^10.4 || ^11.5"
+    "typo3/cms-backend": "^11.5 || ^12.4",
+    "typo3/cms-core": "^11.5 || ^12.4",
+    "typo3/cms-extbase": "^11.5 || ^12.4",
+    "typo3/cms-fluid": "^11.5 || ^12.4",
+    "typo3/cms-fluid-styled-content": "^11.5 || ^12.4",
+    "typo3/cms-frontend": "^11.5 || ^12.4"
   },
   "require-dev": {
-    "helhum/typo3-console": "^6.7 || ^7.1",
+    "bnf/phpstan-psr-container": "^1.0",
+    "friendsoftypo3/phpstan-typo3": "^0.9.0",
+    "helhum/typo3-console": "^7.1",
     "phpstan/phpstan": "^1.8",
     "phpunit/phpunit": "^9.5",
+    "rector/rector": "^0.15",
     "roave/security-advisories": "dev-latest",
+    "slevomat/coding-standard": "^8.8",
     "squizlabs/php_codesniffer": "^3.5",
-    "typo3/minimal": "^10.4 || ^11.5",
+    "typo3/minimal": "^11.5 || ^12.4",
     "typo3/testing-framework": "^6.14",
     "vimeo/psalm": "^5.1"
   },
@@ -55,17 +59,15 @@
     "bin-dir": ".Build/bin",
     "allow-plugins": {
       "typo3/class-alias-loader": true,
-      "typo3/cms-composer-installers": true
+      "typo3/cms-composer-installers": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true
     }
   },
   "scripts": {
-    "post-autoload-dump": [
-      "TYPO3\\TestingFramework\\Composer\\ExtensionTestEnvironment::prepare"
-    ],
     "extension-spreadsheet-fallback": [
       "mkdir -p Resources/Private/Composer",
       "if [ -f Resources/Private/Composer/composer.json ]; then cd Resources/Private/Composer/ && composer update; fi",
-      "if [ ! -f Resources/Private/Composer/composer.json ]; then cd Resources/Private/Composer/ && composer require phpoffice/phpspreadsheet=^1.18 --ignore-platform-reqs && composer config classmap-authoritative true && composer config prepend-autoloader true && composer dump-autoload -o; fi"
+      "if [ ! -f Resources/Private/Composer/composer.json ]; then cd Resources/Private/Composer/ && composer require phpoffice/phpspreadsheet=^1.23 --ignore-platform-reqs && composer config classmap-authoritative true && composer config prepend-autoloader true && composer dump-autoload -o; fi"
     ],
     "prepare-release": [
       "@extension-spreadsheet-fallback",
