@@ -6,14 +6,11 @@ namespace Hoogi91\Spreadsheets\Domain\ValueObject;
 
 use Hoogi91\Spreadsheets\Exception\InvalidDataSourceNameException;
 use JsonSerializable;
+use Stringable;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\MathUtility;
 
-/**
- * Class DsnValueObject
- * @package Hoogi91\Spreadsheets\Domain\ValueObject
- */
-class DsnValueObject implements JsonSerializable
+class DsnValueObject implements JsonSerializable, Stringable
 {
     /**
      * Legacy DSN pattern matches all strings like:
@@ -35,29 +32,15 @@ class DsnValueObject implements JsonSerializable
      */
     private const DSN_PATTERN = '/^spreadsheet:\/\/(\d+)(\?.*)?/';
 
-    /**
-     * @var int
-     */
-    private $fileReference;
+    private int $fileReference = 0;
+
+    private int $sheetIndex = 0;
+
+    private ?string $selection = null;
+
+    private ?string $directionOfSelection = null;
 
     /**
-     * @var int
-     */
-    private $sheetIndex;
-
-    /**
-     * @var string|null
-     */
-    private $selection;
-
-    /**
-     * @var string|null
-     */
-    private $directionOfSelection;
-
-    /**
-     * @param string $dsn
-     *
      * @throws InvalidDataSourceNameException
      */
     public function __construct(string $dsn)
@@ -66,16 +49,21 @@ class DsnValueObject implements JsonSerializable
             $this->legacyDSNParsing($dsn);
         } elseif (preg_match(self::DSN_PATTERN, $dsn) === 1) {
             $dsnData = parse_url($dsn);
-            parse_str($dsnData['query'] ?? '', $queryData);
-            if (MathUtility::canBeInterpretedAsInteger($dsnData['host']) && (int)$dsnData['host'] > 0) {
-                $this->fileReference = (int)$dsnData['host'];
-            } else {
+            if (
+                !isset($dsnData['host'])
+                || !MathUtility::canBeInterpretedAsInteger($dsnData['host'])
+                || (int)$dsnData['host'] <= 0
+            ) {
                 throw new InvalidDataSourceNameException('File reference from DSN is not valid!');
             }
 
-            $this->sheetIndex = (int)($queryData['index'] ?? 0);
-            $this->selection = $queryData['range'] ?? null;
-            $this->directionOfSelection = $queryData['direction'] ?? null;
+            parse_str($dsnData['query'] ?? '', $queryData);
+            $this->fileReference = (int)$dsnData['host'];
+            $this->sheetIndex = MathUtility::canBeInterpretedAsInteger($queryData['index'] ?? null)
+                ? (int)$queryData['index']
+                : 0;
+            $this->selection = is_string($queryData['range'] ?? null) ? $queryData['range'] : null;
+            $this->directionOfSelection = is_string($queryData['direction'] ?? null) ? $queryData['direction'] : null;
         } else {
             throw new InvalidDataSourceNameException('Spreadsheet DSN could not be parsed!');
         }
@@ -85,13 +73,10 @@ class DsnValueObject implements JsonSerializable
         }
     }
 
-    /**
-     * @param string $dsn
-     */
     private function legacyDSNParsing(string $dsn): void
     {
         [$file, $fullSelection] = GeneralUtility::trimExplode('|', $dsn, false, 2);
-        if (strpos($file, 'file:') === 0 && (int)substr($file, 5) !== 0) {
+        if (str_starts_with($file, 'file:') && (int)substr($file, 5) !== 0) {
             $this->fileReference = (int)substr($file, 5);
         } elseif ((int)$file !== 0 && MathUtility::canBeInterpretedAsInteger($file)) {
             $this->fileReference = (int)$file;
@@ -99,23 +84,22 @@ class DsnValueObject implements JsonSerializable
             throw new InvalidDataSourceNameException('File reference from DSN can not be parsed/evaluated!');
         }
 
-        if (trim($fullSelection) !== '') {
-            [$sheetIndex, $selection, $directionOfSelection] = array_pad(
-                GeneralUtility::trimExplode('!', $fullSelection, false, 3),
-                3,
-                ''
-            );
-
-            $this->sheetIndex = (int)($sheetIndex ?: 0);
-            $this->selection = $selection ?: null;
-            $this->directionOfSelection = $directionOfSelection ?: null;
+        if (trim($fullSelection) === '') {
+            return;
         }
+
+        [$sheetIndex, $selection, $directionOfSelection] = array_pad(
+            GeneralUtility::trimExplode('!', $fullSelection, false, 3),
+            3,
+            ''
+        );
+
+        $this->sheetIndex = (int)($sheetIndex ?: 0);
+        $this->selection = $selection ?: null;
+        $this->directionOfSelection = $directionOfSelection ?: null;
     }
 
     /**
-     * @param string $dsn
-     * @return DsnValueObject
-     *
      * @throws InvalidDataSourceNameException
      */
     public static function createFromDSN(string $dsn): DsnValueObject
@@ -123,9 +107,6 @@ class DsnValueObject implements JsonSerializable
         return new self($dsn);
     }
 
-    /**
-     * @return string
-     */
     public function getDsn(): string
     {
         $parameters = [
@@ -136,57 +117,37 @@ class DsnValueObject implements JsonSerializable
 
         $parameters = array_filter(
             $parameters,
-            static function ($value) {
-                return $value !== null;
-            }
+            static fn ($value) => $value !== null
         );
 
         return sprintf('spreadsheet://%d?%s', $this->getFileReference(), http_build_query($parameters));
     }
 
-    /**
-     * @return int
-     */
     public function getFileReference(): int
     {
         return $this->fileReference;
     }
 
-    /**
-     * @return int
-     */
     public function getSheetIndex(): int
     {
         return $this->sheetIndex;
     }
 
-    /**
-     * @return string|null
-     */
     public function getSelection(): ?string
     {
         return $this->selection;
     }
 
-    /**
-     * @return string|null
-     */
     public function getDirectionOfSelection(): ?string
     {
         return $this->directionOfSelection;
     }
 
-    /**
-     * @return string
-     */
-    public function __toString()
+    public function __toString(): string
     {
         return $this->getDsn();
     }
 
-    /**
-     * @return string
-     */
     public function jsonSerialize(): string
     {
         return $this->getDsn();
