@@ -4,41 +4,36 @@ declare(strict_types=1);
 
 namespace Hoogi91\Spreadsheets\Service;
 
+use PhpOffice\PhpSpreadsheet\Cell\CellAddress;
 use PhpOffice\PhpSpreadsheet\Cell\Coordinate;
 use PhpOffice\PhpSpreadsheet\Exception as SpreadsheetException;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 
-/**
- * Class SpanService
- * @package Hoogi91\Spreadsheets\Service
- */
 class SpanService
 {
-
     /**
-     * @var array
+     * @var array<string, array<string>>
      */
-    private static $ignoredColumns = [];
+    private static array $ignoredColumns = [];
 
     /**
-     * @var array
+     * @var array<string, array<int>>
      */
-    private static $ignoredRows = [];
+    private static array $ignoredRows = [];
 
     /**
-     * @var array
+     * @var array<string, array<string>>
      */
-    private static $ignoredCells = [];
+    private static array $ignoredCells = [];
 
     /**
-     * @var array
+     * @var array<string, array<int|string, array<string, int|array<int>>>>
      */
-    private static $mergedCells = [];
+    private static array $mergedCells = [];
 
     /**
-     * @param Worksheet $worksheet
-     * @return array
+     * @return array<string>
      * @throws SpreadsheetException
      */
     public function getIgnoredColumns(Worksheet $worksheet): array
@@ -58,9 +53,7 @@ class SpanService
         // check if unique column values will exceed max. column count and should be ignored
         $maxRowCount = $worksheet->getHighestRow();
         $ignoredColumnsByRow = array_map(
-            static function ($values) use ($maxRowCount) {
-                return count(array_unique($values)) >= $maxRowCount;
-            },
+            static fn ($values) => count(array_unique($values)) >= $maxRowCount,
             $ignoredColumnsByRow
         );
 
@@ -69,8 +62,7 @@ class SpanService
     }
 
     /**
-     * @param Worksheet $worksheet
-     * @return array
+     * @return array<int>
      * @throws SpreadsheetException
      */
     public function getIgnoredRows(Worksheet $worksheet): array
@@ -90,9 +82,7 @@ class SpanService
         // check if unique column values will exceed max. column count and should be ignored
         $maxColumnCount = Coordinate::columnIndexFromString($worksheet->getHighestColumn());
         $ignoredRowsByColumn = array_map(
-            static function ($values) use ($maxColumnCount) {
-                return count(array_unique($values)) >= $maxColumnCount;
-            },
+            static fn ($values) => count(array_unique($values)) >= $maxColumnCount,
             $ignoredRowsByColumn
         );
 
@@ -101,8 +91,7 @@ class SpanService
     }
 
     /**
-     * @param Worksheet $worksheet
-     * @return array
+     * @return array<string>
      */
     public function getIgnoredCells(Worksheet $worksheet): array
     {
@@ -127,8 +116,7 @@ class SpanService
     }
 
     /**
-     * @param Worksheet $worksheet
-     * @return array
+     * @return array<int|string, array<string, int|array<int>>>
      * @throws SpreadsheetException
      */
     public function getMergedCells(Worksheet $worksheet): array
@@ -162,10 +150,8 @@ class SpanService
     }
 
     /**
-     * @param Worksheet $worksheet
-     * @param array $references
-     *
-     * @return array
+     * @param array<array<int>|CellAddress|string> $references
+     * @return array<int>
      */
     private function getCellStyleIndexesFromReferences(Worksheet $worksheet, array $references): array
     {
@@ -175,16 +161,12 @@ class SpanService
                 $result[] = $worksheet->getCell($cellRef)->getXfIndex();
             }
         }
+
         return array_unique($result);
     }
 
-    /**
-     * @param Spreadsheet $spreadsheet
-     * @param Worksheet $worksheet
-     * @return string
-     */
-    private function getActiveSheetHashCode(Spreadsheet $spreadsheet, Worksheet $worksheet): string
+    private function getActiveSheetHashCode(?Spreadsheet $spreadsheet, Worksheet $worksheet): string
     {
-        return md5($spreadsheet->getID() . $worksheet->getHashCode());
+        return md5($spreadsheet?->getID() . $worksheet->getHashCode());
     }
 }
