@@ -1,40 +1,26 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Hoogi91\Spreadsheets\ViewHelpers\Reader;
 
+use InvalidArgumentException;
 use TYPO3\CMS\Core\Resource\FileReference;
 use TYPO3\CMS\Core\Resource\FileRepository;
 use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
-/**
- * Class FileReferenceViewHelper
- * @package Hoogi91\Spreadsheets\ViewHelpers\Reader
- */
 class FileReferenceViewHelper extends AbstractViewHelper
 {
-
-    /**
-     * @var FileRepository
-     */
-    private $fileRepository;
-
-    public function __construct(FileRepository $fileRepository)
+    public function __construct(private readonly FileRepository $fileRepository)
     {
-        $this->fileRepository = $fileRepository;
     }
 
-    /**
-     * Initialize arguments.
-     */
     public function initializeArguments(): void
     {
         parent::initializeArguments();
         $this->registerArgument('uid', 'string', 'File reference uid to resolve', false);
     }
 
-    /**
-     * @return FileReference|null
-     */
     public function render(): ?FileReference
     {
         if (empty($this->arguments['uid'])) {
@@ -43,7 +29,7 @@ class FileReferenceViewHelper extends AbstractViewHelper
 
         try {
             $fileReference = $this->fileRepository->findFileReferenceByUid($this->arguments['uid']);
-        } catch (\InvalidArgumentException $exception) {
+        } catch (InvalidArgumentException) {
             $fileReference = false;
         }
 
