@@ -1,43 +1,33 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Hoogi91\Spreadsheets\Tests\Unit\Service;
 
 use Hoogi91\Spreadsheets\Domain\ValueObject\CellDataValueObject;
 use Hoogi91\Spreadsheets\Domain\ValueObject\DsnValueObject;
 use Hoogi91\Spreadsheets\Exception\InvalidDataSourceNameException;
 use Hoogi91\Spreadsheets\Service;
+use Hoogi91\Spreadsheets\Service\ExtractorService;
 use Hoogi91\Spreadsheets\Tests\Unit\FileRepositoryMockTrait;
 use Hoogi91\Spreadsheets\Tests\Unit\TsfeSetupTrait;
 use PhpOffice\PhpSpreadsheet\Reader\Xlsx;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Http\Message\ServerRequestInterface;
+use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Core\SystemEnvironmentBuilder;
+use TYPO3\CMS\Core\Site\SiteFinder;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
-/**
- * Class ExtractorServiceTest
- * @package Hoogi91\Spreadsheets\Tests\Unit\Service
- */
 class ExtractorServiceTest extends UnitTestCase
 {
     use FileRepositoryMockTrait;
     use TsfeSetupTrait;
 
-    /**
-     * @var Service\ExtractorService
-     */
-    private $extractorService;
+    private ExtractorService $extractorService;
 
-    /**
-     * @var MockObject|Service\ReaderService
-     */
-    private $readerService;
-
-    /**
-     * @var Spreadsheet
-     */
-    private $spreadsheet;
+    private Spreadsheet $spreadsheet;
 
     protected function setUp(): void
     {
@@ -57,9 +47,13 @@ class ExtractorServiceTest extends UnitTestCase
 
         $mappingService = $this->createTestProxy(Service\ValueMappingService::class);
         $styleService = $this->createTestProxy(Service\StyleService::class, [$mappingService]);
+        $cellService = $this->createTestProxy(
+            Service\CellService::class,
+            [$styleService, $this->createMock(SiteFinder::class), $this->createMock(Context::class)]
+        );
         $this->extractorService = new Service\ExtractorService(
             $readerService,
-            $this->createTestProxy(Service\CellService::class, [$styleService]),
+            $cellService,
             $this->createTestProxy(Service\SpanService::class),
             $this->createTestProxy(Service\RangeService::class),
             $mappingService,
@@ -81,7 +75,7 @@ class ExtractorServiceTest extends UnitTestCase
 
     public function testExtractionOfDataByDsnValueObject(): void
     {
-        /** @var MockObject|DsnValueObject $mockDsnValueObject */
+        /** @var MockObject&DsnValueObject $mockDsnValueObject */
         $mockDsnValueObject = $this->getMockBuilder(DsnValueObject::class)->disableOriginalConstructor()->getMock();
         $mockDsnValueObject->expects(self::once())->method('getSheetIndex')->willReturn(0);
         $mockDsnValueObject->expects(self::once())->method('getFileReference')->willReturn(456);
@@ -94,7 +88,7 @@ class ExtractorServiceTest extends UnitTestCase
 
     public function testExtractionOfDataByDsnValueObjectWithRange(): void
     {
-        /** @var MockObject|DsnValueObject $mockDsnValueObject */
+        /** @var MockObject&DsnValueObject $mockDsnValueObject */
         $mockDsnValueObject = $this->getMockBuilder(DsnValueObject::class)->disableOriginalConstructor()->getMock();
         $mockDsnValueObject->expects(self::once())->method('getSheetIndex')->willReturn(0);
         $mockDsnValueObject->expects(self::once())->method('getFileReference')->willReturn(456);
@@ -131,16 +125,13 @@ class ExtractorServiceTest extends UnitTestCase
         $cellValueD5 = $bodyData[5][4];
         self::assertInstanceOf(CellDataValueObject::class, $cellValueD5);
         self::assertEquals(
-            '<span style="color:#000000"><sup>Hoch</sup></span><span style="color:#000000"> Test </span><span style="color:#000000"><sub>Tief</sub></span>',
+            '<span style="color:#000000"><sup>Hoch</sup></span>' .
+            '<span style="color:#000000"> Test </span><span style="color:#000000"><sub>Tief</sub></span>',
             $cellValueD5->getRenderedValue()
         );
     }
 
     /**
-     * @param string $range
-     * @param string $direction
-     * @param bool $cellRef
-     *
      * @dataProvider rangeExtractorDataProvider
      */
     public function testRangeExtractor(string $range, string $direction, bool $cellRef = false): void
@@ -148,28 +139,32 @@ class ExtractorServiceTest extends UnitTestCase
         $worksheet = $this->spreadsheet->getSheet(0);
         $data = $this->extractorService->rangeToCellArray($worksheet, $range, $direction, $cellRef);
 
-        /** @var CellDataValueObject $cellValueA1 */
         if ($direction === Service\ExtractorService::EXTRACT_DIRECTION_HORIZONTAL) {
             $cellValueA1 = $cellRef === false ? $data[1]['B'] : $data[1][2];
         } else {
             $cellValueA1 = $cellRef === false ? $data['B'][1] : $data[2][1];
         }
+        /** @var CellDataValueObject $cellValueA1 */
         self::assertInstanceOf(CellDataValueObject::class, $cellValueA1);
         self::assertEquals('2015', $cellValueA1->getRenderedValue());
 
-        /** @var CellDataValueObject $cellValueE5 */
         if ($direction === Service\ExtractorService::EXTRACT_DIRECTION_HORIZONTAL) {
             $cellValueE5 = $cellRef === false ? $data[5]['D'] : $data[5][4];
         } else {
             $cellValueE5 = $cellRef === false ? $data['D'][5] : $data[4][5];
         }
+        /** @var CellDataValueObject $cellValueE5 */
         self::assertInstanceOf(CellDataValueObject::class, $cellValueE5);
         self::assertEquals(
-            '<span style="color:#000000"><sup>Hoch</sup></span><span style="color:#000000"> Test </span><span style="color:#000000"><sub>Tief</sub></span>',
+            '<span style="color:#000000"><sup>Hoch</sup></span>' .
+            '<span style="color:#000000"> Test </span><span style="color:#000000"><sub>Tief</sub></span>',
             $cellValueE5->getRenderedValue()
         );
     }
 
+    /**
+     * @return array<int, array<string|bool>>
+     */
     public function rangeExtractorDataProvider(): array
     {
         return [
@@ -198,17 +193,17 @@ class ExtractorServiceTest extends UnitTestCase
         self::assertInstanceOf(CellDataValueObject::class, $cellValueA1);
         self::assertEquals('2014', $cellValueA1->getRenderedValue());
 
-        /** @var CellDataValueObject $cellValueA1 */
+        /** @var CellDataValueObject $cellValueC3 */
         $cellValueC3 = $headData[2]['C'];
         self::assertInstanceOf(CellDataValueObject::class, $cellValueC3);
         self::assertEquals('70', $cellValueC3->getRenderedValue());
 
-        /** @var CellDataValueObject $cellValueA1 */
+        /** @var CellDataValueObject $cellValueG3 */
         $cellValueG3 = $bodyData[3]['G'];
         self::assertInstanceOf(CellDataValueObject::class, $cellValueG3);
         self::assertEquals('x', $cellValueG3->getRenderedValue());
 
-        /** @var CellDataValueObject $cellValueA1 */
+        /** @var CellDataValueObject $cellValueA4 */
         $cellValueA4 = $bodyData[4]['A'];
         self::assertInstanceOf(CellDataValueObject::class, $cellValueA4);
         self::assertEquals('©™§∆', $cellValueA4->getRenderedValue());
