@@ -7,19 +7,13 @@ namespace Hoogi91\Spreadsheets\DataProcessing;
 use Hoogi91\Spreadsheets\Domain\ValueObject\DsnValueObject;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 
-/**
- * Class SpreadsheetProcessor
- * @package Hoogi91\Spreadsheets\DataProcessing
- */
 class SpreadsheetProcessor extends AbstractProcessor
 {
-
     /**
-     * Get template relevant data
      * @param DsnValueObject $dsn DSN which is processed
      * @param Spreadsheet $spreadsheet Spreadsheet that is processed
-     * @param array $processedData Processed data
-     * @return array
+     * @param array<string, array<string, int|bool>> $processedData Processed data
+     * @return array<mixed>
      */
     protected function getTemplateData(DsnValueObject $dsn, Spreadsheet $spreadsheet, array $processedData): array
     {
@@ -29,13 +23,14 @@ class SpreadsheetProcessor extends AbstractProcessor
 
         // check if first row of body data should be header
         // 0 = "no header" | 1 = "top" | 2 = "left"
-        $tableHeaderPosition = (int)($processedData['data']['table_header_position'] ?? 0);
+        $tableHeaderPosition = (int) ($processedData['data']['table_header_position'] ?? 0);
         if (empty($headData) && !empty($bodyData) && $tableHeaderPosition === 1) {
             $headData[] = array_shift($bodyData);
         }
 
         // check if last row of body data should be footer
-        $enableTableFooter = (bool)($processedData['data']['table_tfoot'] ?? 0);
+        $footData = [];
+        $enableTableFooter = (bool) ($processedData['data']['table_tfoot'] ?? 0);
         if ($enableTableFooter === true && !empty($bodyData)) {
             $footData[] = array_pop($bodyData);
         }
@@ -45,7 +40,7 @@ class SpreadsheetProcessor extends AbstractProcessor
             'firstColumnIsHeader' => empty($headData) && $tableHeaderPosition === 2,
             'headData' => $headData,
             'bodyData' => $bodyData,
-            'footData' => $footData ?? [],
+            'footData' => $footData,
         ];
     }
 }
