@@ -1,23 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Hoogi91\Spreadsheets\DataProcessing;
 
 use Hoogi91\Spreadsheets\Domain\ValueObject\DsnValueObject;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 
-/**
- * Class TabsProcessor
- * @package Hoogi91\Spreadsheets\DataProcessing
- */
 class TabsProcessor extends AbstractProcessor
 {
-
     /**
-     * Get template relevant data
      * @param DsnValueObject $dsn DSN which is processed
      * @param Spreadsheet $spreadsheet Spreadsheet that is processed
-     * @param array $processedData Processed data
-     * @return array
+     * @param array<mixed> $processedData Processed data
+     * @return array<mixed>
      */
     protected function getTemplateData(DsnValueObject $dsn, Spreadsheet $spreadsheet, array $processedData): array
     {
