@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Hoogi91\Spreadsheets\Tests\Functional\ViewHelpers;
 
 use TYPO3\CMS\Extbase\Object\ObjectManager;
@@ -9,15 +11,25 @@ use TYPO3Fluid\Fluid\View\TemplateView;
 
 abstract class AbstractViewHelperTestCase extends FunctionalTestCase
 {
-
+    /**
+     * @var array<string, non-empty-string>
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
+     */
     protected $pathsToLinkInTestInstance = [
         'typo3conf/ext/spreadsheets/Tests/Fixtures/' => 'fileadmin/user_upload',
     ];
 
+    /**
+     * @var array<non-empty-string>
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
+     */
     protected $testExtensionsToLoad = [
         'typo3conf/ext/spreadsheets',
     ];
 
+    /**
+     * @param array<mixed> $arguments
+     */
     protected function getView(string $template, array $arguments = []): TemplateView
     {
         $view = new TemplateView();
