@@ -1,17 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Hoogi91\Spreadsheets\Tests\Unit;
 
-/**
- * Trait TsfeSetupTrait
- * @package Hoogi91\Spreadsheets\Tests\Unit
- */
+use stdClass;
+
 trait TsfeSetupTrait
 {
-
     public static function setupDefaultTSFE(): void
     {
-        $GLOBALS['TSFE'] = $GLOBALS['TSFE'] ?? new \stdClass();
+        $GLOBALS['TSFE'] ??= new stdClass();
         $GLOBALS['TSFE']->config['config']['locale_all'] = 'de';
     }
 }
